signup
Signup and profile page connecting into child/parent homescreen is done.
Signup only goes to guardian profile as only guardians will sign up in the first place. Log in is not done yet. Also, profile pg is first pg shown as it is eaiser to test. Dummy info is used for mocking purposes for now.